Move encoders to the encoding submodule
- Split the encoding submodule into three parts
1) encodings - contains the underlying functional units
2) loaders - lazy loaders that receive an iterator
3) encoders - classes that are callable and store set parameters
