+++ examples/mnist/minimal_mnist.py
@@ -1,5 +1,5 @@
 from bindsnet.datasets import MNIST
-from bindsnet.encoders import PoissonEncoder
+from bindsnet.encoding import PoissonEncoder
 from bindsnet.pipeline import TorchVisionDatasetPipeline
 from bindsnet.models import DiehlAndCook2015
 from torchvision import transforms
