+++ examples/space_invaders/et_space_invaders.py
@@ -3,7 +3,7 @@ import argparse
 import matplotlib.pyplot as plt
 
 from bindsnet.environment import GymEnvironment
-from bindsnet.encoders import BernoulliEncoder
+from bindsnet.encoding import BernoulliEncoder
 from bindsnet.learning import MSTDPET
 from bindsnet.network import Network
 from bindsnet.network.monitors import Monitor
